+++ lib/sagrone_scraper/agent.rb
@@ -9,12 +9,10 @@ module SagroneScraper
     attr_reader :url, :page
 
     def initialize(options = {})
-      raise Error.new('Exactly one option must be provided: "url" or "page"') unless exactly_one_of(options)
-
-      @url, @page = options[:url], options[:page]
-
-      @url ||= page_url
-      @page ||= http_client.get(url)
+      @url = options.fetch(:url) do
+              raise Error.new('Option "url" must be provided')
+             end
+      @page = http_client.get(url)
     rescue StandardError => error
       raise Error.new(error.message)
     end
@@ -28,19 +26,6 @@ module SagroneScraper
         agent.user_agent_alias = AGENT_ALIASES.sample
         agent.max_history = 0
       end
-    end
-
-    private
-
-    def page_url
-      @page.uri.to_s
-    end
-
-    def exactly_one_of(options)
-      url_present = !!options[:url]
-      page_present = !!options[:page]
-
-      (url_present && !page_present) || (!url_present && page_present)
     end
   end
 end