+++ lib/sagrone_scraper/base.rb
@@ -5,13 +5,15 @@ module SagroneScraper
   class Base
     Error = Class.new(RuntimeError)
 
-    attr_reader :page, :page_url, :attributes
+    attr_reader :page, :url, :attributes
 
     def initialize(options = {})
-      @page = options.fetch(:page) do
-                raise Error.new('Option "page" must be provided.')
-              end
-      @page_url = @page.uri.to_s
+      raise Error.new('Exactly one option must be provided: "url" or "page"') unless exactly_one_of(options)
+
+      @url, @page = options[:url], options[:page]
+
+      @url ||= page_url
+      @page ||= Agent.new(url: url).page
       @attributes = {}
     end
 
@@ -34,6 +36,17 @@ module SagroneScraper
     end
 
     private
+
+    def exactly_one_of(options)
+      url_present = !!options[:url]
+      page_present = !!options[:page]
+
+      (url_present && !page_present) || (!url_present && page_present)
+    end
+
+    def page_url
+      @page.uri.to_s
+    end
 
     def self.method_names
       @method_names ||= []